Refactor ledger tests, introduce ledger mocking for tests
