+++ crypto/knownValues.go
@@ -1,0 +1,6 @@
+package crypto
+
+const (
+	// Tests expect a ledger device initialized to the following mnemonic
+	TestMnemonic = "equip will roof matter pink blind book anxiety banner elbow sun young"
+)
